moved image banner from menu to about touhoufest page
